+++ make.py
@@ -156,16 +156,6 @@ def ut():
 def ft():
     run_cmd('py.test -k "functional" -v --strict')
 
-
-@task('run all v12 unit tests under current interpreter, and print coverage report')
-def v12_ut():
-    run_cmd('coverage run --source bingads -m py.test v12tests/ -k "not functional" -v --strict')
-    run_cmd('coverage report')
-	
-@task('run all v12 functional tests under current interpreter.')
-def v12_ft():
-    run_cmd('py.test v12tests/ -k "functional" -v --strict')
-
 @task('run all v13 unit tests under current interpreter, and print coverage report')
 def v13_ut():
     run_cmd('coverage run --source bingads -m py.test v13tests/ -k "not functional" -v --strict')